+++ docs/changelog.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## v7.0.5 / v0.1.3 (sdk) - _2023-09-20_
+
+### Fixes
+
+- Fix the behaviour of the dataset pattern functions in the SDK.
+
+## v7.0.4 / v0.1.2 (sdk) - _2023-09-20_
+
+### Features
+
+- Improved release process
+- Added Athena workgroup and database as outputs of the rAPId module.
+
+### Fixes
+
+- Updated terraform default `application_version` and `ui_version` variables.
+- Migration script and documentation.
+
 ## v7.0.3 / v0.1.2 (sdk) - _2023-09-15_
 
 ### Fixes
@@ -39,7 +57,8 @@
 
 - See the [migration doc](migration.md) for details on how to migrate to v7 from v6.
 
-[Unreleased changes]: https://github.com/no10ds/rapid/compare/v7.0.3...HEAD
+[Unreleased changes]: https://github.com/no10ds/rapid/compare/v7.0.4...HEAD
+[v7.0.4 / v0.1.2 (sdk)]: https://github.com/no10ds/rapid/v7.0.3...v7.0.4
 [v7.0.3 / v0.1.2 (sdk)]: https://github.com/no10ds/rapid/v7.0.2...v7.0.3
 [v7.0.2 / v0.1.2 (sdk)]: https://github.com/no10ds/rapid/v7.0.1...v7.0.2
 [v7.0.1 / v0.1.2 (sdk)]: https://github.com/no10ds/rapid/v7.0.0...v7.0.1